requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3.0   # only if you want JWT option
django-cors-headers>=4.3
django-environ>=0.11
whitenoise>=6.6.0                      # for static files (production)
gunicorn>=22.0.0                       # if deploying on Linux
psycopg2-binary>=2.9                   # if using PostgreSQL
mysqlclient>=2.2                       # (optional, if using MySQL instead)
